backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
cachetools==5.5.2
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.5
cryptography==45.0.6
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.57.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
httplib2==0.30.0
idna==3.10
kombu==5.5.4
lxml==6.0.1
oauthlib==3.3.1
packaging==25.0
Pillow==10.1.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-magic==0.4.27
python-pptx==1.0.2
python3-openid==3.2.0
pytz==2025.2
redis==5.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
xlsxwriter==3.2.5
